+++ rector.php
@@ -24,6 +24,8 @@ return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->paths([
         __DIR__ . '/src',
         __DIR__ . '/tests',
+        __DIR__ . '/symfony-bridge/src',
+        __DIR__ . '/symfony-bridge/tests',
     ]);
 
     $rectorConfig->skip([